requirements.txt

numpy==1.26.4
Pillow==10.2.0
protobuf==3.20.3
tensorflow-cpu==2.15.0
keras==2.15.0

fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
ml-dtypes==0.2.0

streamlit==1.33.0
matplotlib==3.8.3
requests==2.31.0